Time: 8 ms (53.21%), Space: 43.5 MB (79.38%) - LeetHub

+++ 1458-sort-integers-by-the-number-of-1-bits/1458-sort-integers-by-the-number-of-1-bits.java
@@ -1,10 +1,4 @@
-
 class Solution {
-    // class Cmp implements Comparator<Integer>{
-    //     public int compare(Integer a, Integer b){
-
-    //     }
-    // }
     public int[] sortByBits(int[] arr) {
         int n=arr.length;
         List<Integer>al=new ArrayList<>();
@@ -19,7 +13,8 @@ class Solution {
         });
         int ans[]=new int[n]; int k=0;
         for(Integer x:al){
-            ans[k++]=(int)x;
+            // ans[k++]=(int)x;
+            ans[k++]=x;
         }
         return ans;
     }
